Storage click behavior changes (#2500)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
- Right clicking storage items with an empty hand has the same behavior
as left click
- Right clicking storage with an item does not put it in
- Restored right clicking a sack to take a random thing out 
- **You can no longer right click keyrings and bins to open them.
Instead do mousedrag or alt click**
- **You can't alt click to open backpacks anymore from your back.**

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Keyrings and bins were changed to remove snowflake code. Complain to me
if you really want it back.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->