Deep speak, Fish Tongue & Triton Accent (#2561)
## About The Pull Request

Adds Deep Speak, a language for Triton with it's own font, and
Scandinavian syllables.

Repurposes the unused lizard tongue into a fish tongue that changes
spoken s & z in to sss & zzz. Gives it a sprite.

Adds an accent for Triton. Mostly based off Norwegian/Scandinavian,
please looks at this and ask for adjustments. BCBM2222 on the discord
asked for this, and sent in a file that I trimmed down and adjusted.
Honestly, the tongue and the accent might be a bit too much. Let me know
if we need to cut it. So let me know for adjustments.

## Why It's Good For The Game

The Tritons seem like they would have their own language. Accent might
be a bit much, but that and their new tongues help make them even more
unique.
<img width="2560" height="1440" alt="image"
src="https://github.com/user-attachments/assets/3318c850-32f0-4e5d-83da-226aa974a475"
/>


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->