Lore Accurate Arcyne Jobs (#2480)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Puts the acryne into a more lore accurate space. Removal of Sorceress,
tweaking of court magician, ~~addition of a mage associate (Concept
borrowed from Azure Peak, but with tweaks)~~ (removed per request), and
a overhaul of the job descriptions.
## Why It's Good For The Game

Aberra wanted this

But currently jobs doing Arcyne Magick need a *desperate* loreification.
They reference outdated lore, or contradict how the arcyne works in the
setting. As such, this PR does a pass to 1. make sure it's up to lore,
and 2. prepare for further changes to make magick more lore accurate
(and to balance it.)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: PotatoTomahto <76505086+PotatoTomahto@users.noreply.github.com>