[Draft] Vampires reworked (#2384)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Using WOD13 as a baseline I went over vampires and reworked them, uses
the ideas of clans and disciplines from WOD13 and VtM in general but the
actual implmentation is vastly different

This also turned into a merchantry rework, base design for traders came
from:
https://github.com/tgstation/tgstation/pull/79187

Ports the portable table and stove from with code cleanup and vanderlin
specific changes:
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/3084

ports field find_targets, redoes it for find_and_set
https://github.com/tgstation/tgstation/pull/82640

Modified to fit world factions
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->