Ascendant effects for gods interventions (#2612)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR adds additional effects for gods' interventions when the
relevant god is ascendant.

List of effects:

- Astrata - Buff is now applied to all astrata's followers, not just
clergy and nobles
- Noc - Even cheaper magic
- Malum - Even better mining yields
- Abyssor - Even deadlier drowning
- Baotha - Alcohol last even longer
- Matthios - One more skill level is applied when pickpocketing
- Graggar - It will be always grand culling instead of it being random
- Necra - All current deadites will lose their zombie status and die
- Eora - Even more couples will be matchmaked
- Xylix - Fortune change is now rolled for everyone, not just his
followers
- Zizo - All currently dead people will immediately rise as deadites
- Dendor - Fertility and drought traits are stronger on crops, blessed
and malevont vines are more plentiful
- Ravox - Weakest follower gets bonus to all stats instead of just
combat ones
- Pestra - Corpses take even longer to rot

Also makes a define for patron/storyteller names to prevent mistakes, as
they are quite interconnected now.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

More interesting ascendant rounds.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl: Arkatos
add: Added ascendant effects for gods' interventions, which makes them
stronger and more impactful when the relevant god is ascendant.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->