Hollow Ranger Fixes (#2712)
## About The Pull Request

Updates hollowranger.dm's allowed_races to be in line with the other
adventurer classes so it can actually spawn without triumph buying, also
gives it some combat music because it didn't have any specified. This is
my first time messing around with github so yell at me if I did
something wrong.

## Why It's Good For The Game

An interesting adventurer class was completely unable to spawn because
of the way the species restrictions were entered.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.