Slight balance changes for bodypart wounds and crits (#2505)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Increase the thresholds of all but biting and lashing to:
30 to INF for severe
15 to 30 for normal
5 to 15 for weak
This is because almost everything immediately does the highest tier of
wound, even goblins with big sticks or daggers. lashing and biting
exempt because lashing is a joke and biting also is outside of
werewolves, vampires and female tritons.

Perhaps controversial:
Biting cannot crit unless its a severe bite as wounds before that do not
cause bleeding thus cutting arteries doesn't make sense.

Getting a crit replaces the wound that would've been applied by that
attack.

Causing a crit requires a wound to be in queue to be applied in the
first place, this mostly just means attacks with no class or attacks
less than 5 damage can't go into the crit proc which returns early in
these cases anyway.

Fix chest and head wounds not having the 5 damage threshold.

These changes do not effect simple wounds, I don't event think they work
properly.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently almost every wound you apply on someone that does meaningful
damage (isn't blocked by armour) does the highest tier of wound, why
have tiers if the most common weapons by default have 20 force or
greater?

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->