Gives Hollow-kin Beastish-speak. (#2744)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Gives Hollow-kin access to the beast language that Werewolves use.
**Polyglot special and Archivist** are exempt from this! They do not
know Beastish/do not bother learning it.

## Why It's Good For The Game

Adds credence to the idea that Hollowkin are harbingers of Lycanthropy
and gives them a unique language that further separates them from the
Humen species.

<img width="1012" height="653" alt="image"
src="https://github.com/user-attachments/assets/a68dd8bf-68fd-4c5e-aab8-a1ae26719b79"
/>


## Changelog

:cl: Wabberspace
add: Added Beastish to hollow-kin/demihumens.
/:cl:
<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->