Makes Churn Undead Better Than Abrogation (#2688)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Churn undead was in a pitiful state, Abrogation was clearly better both
in cost, damage and giving the undead unbuffs, giving around 91 damage
and -2 to all STATs for the entire duration also making them shine
making them stand apart.


![totaldmgaborro](https://github.com/user-attachments/assets/30ff2a16-01a4-4742-b8e6-c47978575afc)

While Churn Undead stunned for abit but only did 20-40 DMG with a 80%
chance of failure for a Templar with their start miracle skill being
Average so wasting 60 devotion and getting a cd of 35:


![0_tile_damage](https://github.com/user-attachments/assets/32965cfe-9d5d-4cd3-b352-a2bf70169729)

This PR aims to balance it out making it stronger than abrogation with a
few buffs and debuffs, first of all the cooldown is increased to 60
seconds and cost to 100 but Templars with Average miracle got a 40%
chance of hitting the undead with it, with expert it is a 80% and if
they reach Master or above it is a 100%. (Templar hitchance with average
used to be 20%)

Churn undead now also throw the skeletons far from the Templar/Acolyte
who uses the ability, setting them on fire doing the same damage as a
lesser miracle on an undead, stunning them for double the time, paralyze
is just so they fall, doing around 80-100 damage but it could be
increased if the undead hit eachother or trees,rocks,walls etc, this
could also make them fall into the river or from high places.
The ability of gibbing a poltergeist on usage was also added back, i
didn't saw any clear reason why it should go, necra is not fond of
undead nor ghost who become evil and attack the living.

New: If the Vampire Lord is not ascended, Churn undead also hurts them


## Why It's Good For The Game

Makes Churn Undead useful as it is the ultimate skill from Necra
Acolytes and the only offensive skill of Necra Templars.

## Changelog

:cl:
balance: Buff Churn Undead
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>