New Migrant wave : Magic School Expedition (#2808)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

A new migrant wave that goes around a teacher and abunch of students
that went into an expedition on foot from Kingsfield to Vanderlin Magic
Guild, teacher is a good mage, studens are half decent the upsides is
that they are not broke as the mage adventurer with pouches with stuff,
the mana necklace and better books, still they are children and since it
is rare they won't get many nights to get too strong.

## Why It's Good For The Game

New migrant wave that i think that could be fun, specially for those
that want to try out mage stuff without going through the chore of
getting a better start book and find the resources to summon elementals.

The teacher and children dynamic could also have some good rp.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: A new migrant wave called " Magic School Expedition"
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->