Another Maniac Addition (#2823)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR give maniac strong grabber, reduce the role restrictions due to
the fact that youngling get nerfed when they're a maniac and increase
the chance of midround maniac proccing(for your information, it still
has NEVER rolled yet.), also re-add the special examine for maniac.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More maniac changes good, also like potato would say, the examine is
sovl. on another note, maniac is just a one man only antag, it should be
able to get it's ground, and i wasn't aware that assassin was able to
receive strong grab trait, which is something that maniac desperately
needed.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Re-add the maniac's disgusting pig examine.
balance: increase the weight and chance for it to proc of the midround
maniac proccing(it still hasn't rolled yet in ages.)
add: Give Maniac strong-grabber trait
balance: lower maniac's role restrictions(mostly youngling roles since
they do get debuffed)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>
Co-authored-by: VANDERLIN-MONARCH <224467185+VANDERLIN-MONARCH@users.noreply.github.com>
Co-authored-by: PotatoTomahto <76505086+PotatoTomahto@users.noreply.github.com>